CI stuff (#1)
* Travis shit work in progress

* CI shit work in progress

* CI shit work in progress

* CI + workin Auto-update minimal project branch for future reference
It doesn't include the signing steps yet, so that needs to be figured out in the future

* Connected electron-builder to react-rca for dev server and prod releases

* Some cleaning and refactoring

* Adding some stupid semicolon to satisfy the linter shit :D

* Writing and Testing the travis yaml
This takes a lot of pushes and trials so we'll create a new branch and get rid of it later on after rebasing / squashing everything to another branch! ;)

* Writing and Testing the travis yaml
This takes a lot of pushes and trials so we'll create a new branch and get rid of it later on after rebasing / squashing everything to another branch! ;)

* Writing and Testing the travis yaml
This takes a lot of pushes and trials so we'll create a new branch and get rid of it later on after rebasing / squashing everything to another branch! ;)

* Writing and Testing the travis yaml
This takes a lot of pushes and trials so we'll create a new branch and get rid of it later on after rebasing / squashing everything to another branch! ;)

* Writing and Testing the travis yaml
This takes a lot of pushes and trials so we'll create a new branch and get rid of it later on after rebasing / squashing everything to another branch! ;)

* Writing and Testing the travis yaml
This takes a lot of pushes and trials so we'll create a new branch and get rid of it later on after rebasing / squashing everything to another branch! ;)

* Writing and Testing the travis yaml
This takes a lot of pushes and trials so we'll create a new branch and get rid of it later on after rebasing / squashing everything to another branch! ;)

* Writing and Testing the travis yaml
This takes a lot of pushes and trials so we'll create a new branch and get rid of it later on after rebasing / squashing everything to another branch! ;)

* Writing and Testing the travis yaml
This takes a lot of pushes and trials so we'll create a new branch and get rid of it later on after rebasing / squashing everything to another branch! ;)

* Writing and Testing the travis yaml
This takes a lot of pushes and trials so we'll create a new branch and get rid of it later on after rebasing / squashing everything to another branch! ;)

* Writing and Testing the travis yaml
This takes a lot of pushes and trials so we'll create a new branch and get rid of it later on after rebasing / squashing everything to another branch! ;)

* Writing and Testing the travis yaml
This takes a lot of pushes and trials so we'll create a new branch and get rid of it later on after rebasing / squashing everything to another branch! ;)

* Writing and Testing the travis yaml ( release candidate )
This takes a lot of pushes and trials so we'll create a new branch and get rid of it later on after rebasing / squashing everything to another branch! ;)

* Writing and Testing the travis yaml ( release candidate )
This takes a lot of pushes and trials so we'll create a new branch and get rid of it later on after rebasing / squashing everything to another branch! ;)

* Writing and Testing the travis yaml ( release candidate )
This takes a lot of pushes and trials so we'll create a new branch and get rid of it later on after rebasing / squashing everything to another branch! ;)

* Writing and Testing the travis yaml ( release candidate )
This takes a lot of pushes and trials so we'll create a new branch and get rid of it later on after rebasing / squashing everything to another branch! ;)

* Writing and Testing the travis yaml ( release candidate )
This takes a lot of pushes and trials so we'll create a new branch and get rid of it later on after rebasing / squashing everything to another branch! ;)

* Writing and Testing the travis yaml ( release candidate )
This takes a lot of pushes and trials so we'll create a new branch and get rid of it later on after rebasing / squashing everything to another branch! ;)

* Writing and Testing the travis yaml ( release ) [skip travis]
This takes a lot of pushes and trials so we'll create a new branch and get rid of it later on after rebasing / squashing everything to another branch! ;)

* Writing and Testing the travis yaml ( release ) [skip travis]
This takes a lot of pushes and trials so we'll create a new branch and get rid of it later on after rebasing / squashing everything to another branch! ;)

* Writing and Testing the travis yaml ( release ) [skip ci]
This takes a lot of pushes and trials so we'll create a new branch and get rid of it later on after rebasing / squashing everything to another branch! ;)

* Writing and Testing the travis yaml ( release )
This takes a lot of pushes and trials so we'll create a new branch and get rid of it later on after rebasing / squashing everything to another branch! ;)

* Writing and Testing the travis yaml ( release )
This takes a lot of pushes and trials so we'll create a new branch and get rid of it later on after rebasing / squashing everything to another branch! ;)

* Writing and Testing the travis yaml ( release candidate )
This takes a lot of pushes and trials so we'll create a new branch and get rid of it later on after rebasing / squashing everything to another branch! ;)

* Writing and Testing the travis yaml ( release candidate )
This takes a lot of pushes and trials so we'll create a new branch and get rid of it later on after rebasing / squashing everything to another branch! ;)

* Writing and Testing the travis yaml ( release candidate )
This takes a lot of pushes and trials so we'll create a new branch and get rid of it later on after rebasing / squashing everything to another branch! ;)

* Writing and Testing the travis yaml ( release candidate )
This takes a lot of pushes and trials so we'll create a new branch and get rid of it later on after rebasing / squashing everything to another branch! ;)

* Writing and Testing the travis yaml ( release candidate )
This takes a lot of pushes and trials so we'll create a new branch and get rid of it later on after rebasing / squashing everything to another branch! ;)

* [skip ci]